add test for empty item name
